services/studio/app/google_search_agent/requirements.txt
google-cloud-aiplatform[agent_engines,adk]>=1.112.0
google-cloud-storage>=3.0.0
opentelemetry-instrumentation-google-genai>=0.1b0
websocket-client>=15.0.1
google-genai==1.56.0
websockets==15.0.1
python-dotenv==1.2.1
fastapi==0.128.3
uvicorn==0.40.0
google-adk==1.24.1